+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Unless otherwise specified, any version comparison below is the comparison of se
 ### Added
 
 - Implemented trapdoor except redstone feature (Redstone feature requires the implementation of redstone system).
+- Implemented sponge and wet sponge.
 - Introduced [sentry](https://www.sentry.io) to capture exception and upload them to sentry server automatically, which
   helps us to track and fix bug more efficiently. Sentry is only enabled in non-dev version.